Port project from java 8, netbeans 8.1 to a Modular Java 11 project using NetBeans 11.0, jdk-11.0.4 & javafx-sdk-11.0.2 using branch java11 until ready to be merged into master. Exclude dist/ (custom runtime image) as libjvm.so (297MB) exceeds github file size limit of 100MB. Use github Releases and attach a file containing compressed dist/... to the release as there is no file size limitation on that.
