+++ build.xml
@@ -7,8 +7,8 @@
 <!-- the Compile on Save feature is turned off for the project. -->
 <!-- You can turn off the Compile on Save (or Deploy on Save) setting -->
 <!-- in the project's Project Properties dialog box.-->
-<project name="BackupGnuCash" default="default" basedir=".">
-    <description>Builds, tests, and runs the project BackupGnuCash.</description>
+<project name="BackupGnuCashJ11" default="default" basedir="." xmlns:j2semodularproject="http://www.netbeans.org/ns/j2se-modular-project/1">
+    <description>Builds, tests, and runs the project BackupGnuCashJ11.</description>
     <import file="nbproject/build-impl.xml"/>
     <!--
 
@@ -55,14 +55,6 @@
       run:                      execution of project 
       -javadoc-build:           Javadoc generation
       test-report:              JUnit report generation
-
-    An example of overriding the target for project execution could look like this:
-
-        <target name="run" depends="BackupGnuCash-impl.jar">
-            <exec dir="bin" executable="launcher.exe">
-                <arg file="${dist.jar}"/>
-            </exec>
-        </target>
 
     Notice that the overridden target depends on the jar target and not only on 
     the compile target as the regular run target does. Again, for a list of available
