+++ nbproject/private/private.properties
@@ -1,6 +1,8 @@
 compile.on.save=true
 do.depend=false
 do.jar=true
+do.jlink=true
 javac.debug=true
 javadoc.preview=true
-user.properties.file=/home/cgood/.netbeans/8.1/build.properties
+jlink.strip=false
+user.properties.file=/home/cgood/.netbeans/11.0/build.properties
