+++ nbproject/private/private.xml
@@ -1,19 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="1">
-        <file>
-            <url>src/backupgnucash/BackupGnuCashController.java</url>
-            <bookmark id="1">
-                <name/>
-                <line>134</line>
-                <key/>
-            </bookmark>
-        </file>
-    </editor-bookmarks>
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
         <group>
-            <file>file:/home/cgood/NetBeansProjects/BackupGnuCash/src/backupgnucash/BackupGnuCashController.java</file>
-            <file>file:/home/cgood/NetBeansProjects/BackupGnuCash/README.md</file>
+            <file>file:/home/cgood/NetBeansProjects/BackupGnuCashJ11/src/backupgnucash/classes/module-info.java</file>
+            <file>file:/home/cgood/NetBeansProjects/BackupGnuCashJ11/src/backupgnucash/classes/org/openjfx/BackupGnuCashController.java</file>
         </group>
     </open-files>
 </project-private>
