+++ nbproject/project.properties
@@ -3,7 +3,8 @@ annotation.processing.enabled.in.editor=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=BackupGnuCash
+application.desc=Backup GnuCash data and configuration files.
+application.title=BackupGnuCashJ11
 application.vendor=cgood
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
@@ -11,45 +12,56 @@ build.classes.excludes=**/*.java,**/*.form
 build.dir=build
 build.generated.dir=${build.dir}/generated
 build.generated.sources.dir=${build.dir}/generated-sources
+build.modules.dir=${build.dir}/modules
 # Only compile against the classpath explicitly listed here:
 build.sysclasspath=ignore
 build.test.classes.dir=${build.dir}/test/classes
+build.test.modules.dir=${build.dir}/test/modules
 build.test.results.dir=${build.dir}/test/results
 # Uncomment to specify the preferred debugger connection transport:
 #debug.transport=dt_socket
 debug.classpath=\
     ${run.classpath}
+debug.modulepath=\
+    ${run.modulepath}
 debug.test.classpath=\
     ${run.test.classpath}
+debug.test.modulepath=\
+    ${run.test.modulepath}
 # Files in build.classes.dir which should be excluded from distribution jar
 dist.archive.excludes=
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/BackupGnuCash.jar
 dist.javadoc.dir=${dist.dir}/javadoc
+dist.jlink.dir=${dist.dir}/jlink
+dist.jlink.output=${dist.jlink.dir}/BackupGnuCashJ11
 endorsed.classpath=
 excludes=
 includes=**
-jar.archive.disabled=${jnlp.enabled}
 jar.compress=false
-jar.index=${jnlp.enabled}
 javac.classpath=
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
-javac.external.vm=true
+javac.external.vm=false
+javac.modulepath=\
+    ${libs.JavaFX11.classpath}
+javac.processormodulepath=
 javac.processorpath=\
     ${javac.classpath}
-javac.source=1.8
-javac.target=1.8
+javac.source=11
+javac.target=11
 javac.test.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}
+    ${javac.classpath}
+javac.test.modulepath=\
+    ${javac.modulepath}:\
+    ${build.modules.dir}
 javac.test.processorpath=\
     ${javac.test.classpath}
 javadoc.additionalparam=
 javadoc.author=false
 javadoc.encoding=${source.encoding}
+javadoc.html5=false
 javadoc.noindex=false
 javadoc.nonavbar=false
 javadoc.notree=false
@@ -58,38 +70,31 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-jnlp.codebase.type=no.codebase
-jnlp.descriptor=application
-jnlp.enabled=false
-jnlp.mixed.code=default
-jnlp.offline-allowed=false
-jnlp.signed=false
-jnlp.signing=
-jnlp.signing.alias=
-jnlp.signing.keystore=
-main.class=backupgnucash.BackupGnuCash
-# Optional override of default Application-Library-Allowable-Codebase attribute identifying the locations where your signed RIA is expected to be found.
-manifest.custom.application.library.allowable.codebase=
-# Optional override of default Caller-Allowable-Codebase attribute identifying the domains from which JavaScript code can make calls to your RIA without security prompts.
-manifest.custom.caller.allowable.codebase=
-# Optional override of default Codebase manifest attribute, use to prevent RIAs from being repurposed
-manifest.custom.codebase=
-# Optional override of default Permissions manifest attribute (supported values: sandbox, all-permissions)
-manifest.custom.permissions=
-manifest.file=manifest.mf
-meta.inf.dir=${src.dir}/META-INF
-mkdist.disabled=false
+# The jlink additional root modules to resolve
+jlink.additionalmodules=
+# The jlink additional command line parameters
+jlink.additionalparam=
+jlink.launcher=true
+jlink.launcher.name=BackupGnuCash
+main.class=org.openjfx.BackupGnuCash
 platform.active=default_platform
-run.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}
+project.license=gpl30
+run.classpath=
 # Space-separated list of JVM arguments used when running the project.
 # You may also define separate properties like run-sys-prop.name=value instead of -Dname=value.
 # To set system properties for unit tests define test-sys-prop.name=value:
 run.jvmargs=
+run.modulepath=\
+    ${libs.JavaFXMODS11.classpath}:\
+    ${javac.modulepath}:\
+    ${build.modules.dir}
 run.test.classpath=\
-    ${javac.test.classpath}:\
-    ${build.test.classes.dir}
+    ${javac.test.classpath}
+run.test.modulepath=\
+    ${javac.test.modulepath}:\
+    ${build.test.modules.dir}
 source.encoding=UTF-8
 src.dir=src
-test.src.dir=test
+src.dir.path=classes
+test.src.dir=src
+test.src.dir.path=tests
