+++ src/backupgnucash/classes/org/openjfx/BackupGnuCash.css
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 Chris Good
+ * Copyright (C) 2019 Chris Good
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
